Fix issues with draft editing antique licenses and files on new drafts
Also resolves an issue with autoloading conflicting with the cache reloader in some cases
