+++ app/models/jupiter_core.rb
@@ -1,0 +1,16 @@
+module JupiterCore
+  extend ActiveSupport::Autoload
+
+  class ObjectNotFound < StandardError; end
+  class PropertyInvalidError < StandardError; end
+  class MultipleIdViolationError < StandardError; end
+  class AlreadyDefinedError < StandardError; end
+  class LockedInstanceError < StandardError; end
+
+  VISIBILITY_PUBLIC = CONTROLLED_VOCABULARIES[:visibility].public.freeze
+  VISIBILITY_PRIVATE = CONTROLLED_VOCABULARIES[:visibility].private.freeze
+  VISIBILITY_AUTHENTICATED = CONTROLLED_VOCABULARIES[:visibility].authenticated.freeze
+
+  VISIBILITIES = [VISIBILITY_PUBLIC, VISIBILITY_PRIVATE, VISIBILITY_AUTHENTICATED].freeze
+
+end
