+++ app/models/jupiter_core/locked_ldp_object.rb
@@ -1,17 +1,9 @@
 # JupiterCore::LockedLdpObject classes are lightweight, read-only objects
 # TODO: this file could benefit from some reorganization, possibly into several files
 module JupiterCore
-  class ObjectNotFound < StandardError; end
-  class PropertyInvalidError < StandardError; end
-  class MultipleIdViolationError < StandardError; end
-  class AlreadyDefinedError < StandardError; end
-  class LockedInstanceError < StandardError; end
 
-  VISIBILITY_PUBLIC = CONTROLLED_VOCABULARIES[:visibility].public.freeze
-  VISIBILITY_PRIVATE = CONTROLLED_VOCABULARIES[:visibility].private.freeze
-  VISIBILITY_AUTHENTICATED = CONTROLLED_VOCABULARIES[:visibility].authenticated.freeze
 
-  VISIBILITIES = [VISIBILITY_PUBLIC, VISIBILITY_PRIVATE, VISIBILITY_AUTHENTICATED].freeze
+
 
   class LockedLdpObject
